Allow one parent folder for distribution
Before this change, all FAW distributions needed to have their code in
the same folder as the FAW repository. For private code, this was a
complete hassle. However, Docker disallows accessing parent directories.
Therefore, `workbench.py` has been modified to support a
one-level-parent-traversal to a distribution folder. It will double the
storage size on disk, as the Docker images for FAW-official and private
distributions will be different, but should be a tenable solution.