Optimization for long file names ... fixes Chrome on 300k
Note that the error caused by this is very difficult to detect..
basically only visible because the websocket fails to parse the JSON
message (it gets truncated). It's not really a fix to break this message
up into chunks, because we're also butting up against the per-tab memory
limit in e.g. Chrome. Best way forward is to subsample files, I would
think.