Clustering plugin optimization / retry on error screen
Rather than running clustering on full N^2 pairwise computation, only
run for a given feature of interest. Increases load time per feature,
but greatly reduces runtime for a large number of features (as it brings
it down to O(N) runtime for each feature).

Previously, the UI for the FAW would show an error in the IFrame for
decision plugins on failure. It turns out, particularly when debugging
JSON arguments, this was annoying. So, added a retry button to ease
debugging of such invocations.