Move common components related to processing decisions to a different shared subfolder.
This includes a number of common shared data structures, the DSL parser as well as the
decision processing code itself. Many other files have been touched to fix issues related
to this move.

Also for simplicity, the common folder is not a package we separately install, but just
a folder that is symlinked appropriately.