Add a "front-end" to FAW for use with CI (#82)
Unified frontend and backend processing of decisions, and adds a web hook for validating decisions.

* Pull the bulk of `reprocess` in to a separate file.

This is the first step of refactoring `reprocess` and related aspects in to a shared module/directory to support reuse across the UI and CI aspects of FAW.

* Move common components related to processing decisions to a different shared subfolder.

This includes a number of common shared data structures, the DSL parser as well as the
decision processing code itself. Many other files have been touched to fix issues related
to this move.

Also for simplicity, the common folder is not a package we separately install, but just
a folder that is symlinked appropriately.

* File filter processing moved back to the UI

* Refactor `api_decisions_get` and related functions.

`api_decisions_get` and `_statsbyfile_cursor` now have their common functionality extracted out in to separate functions, to better support the CI front-end work, where we will not have access to `vuespa` client instances.

* Add an endpoint to the FAW server to support querying decisions.

An additional endpoint `/decisions` has been added to the FAW
webserver. The endpoint takes two parameters:
  - `analysis_set_id`: the id of the analysis set for which to return decisions.
  - `dsl`: the (text for) the decision dsl to use to (well) make decisions.

The decisions are returned as a JSON list.

* Add an endpoint to the CI server to fetch decisions (delegating to the FAW server).

* Support for fetching decisions in FAW CI command line tool.

* Fixes for CI build process + turning decision harvesting to asynchronous

* Minor cleanup

---------

Co-authored-by: Walt Woods <waltw@galois.com>