Fix for disabling a parser mid-parse
Before this, the parsers array could have a key that had no version,
which would crash when that version was fetched during the normal parse
pipeline. Now, the dispatcher validates versions before sending them
out... There is probably a more efficient way to do this, but this
should at least be correct, in that any version change / pipeline change
would re-trigger the analysis set build stage, causing the correct
version to be written.